libs/aiotraq/pyproject.toml
[tool.poetry]
name = "aiotraq"
version = "3.0"
description = "Async ready traQ API Client for Python"
authors = []
readme = "README.md"
packages = [
    {include = "aiotraq"},
]
include = ["CHANGELOG.md", "aiotraq/py.typed"]


[tool.poetry.dependencies]
python = "^3.8"
httpx = ">=0.20.0,<0.28.0"
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
